+++ src/api.js
@@ -18,3 +18,12 @@ export function useWallet (newWallet) {
   factorsManager.useApi(api)
   walletsManager.useApi(api)
 }
+
+export async function loadingDataViaLoop (response) {
+  let data = response.data
+  while (response.data.length) {
+    response = await response.fetchNext()
+    data = [...data, ...response.data]
+  }
+  return data
+}