+++ src/vue/modules/dividend-form/index.vue
@@ -179,7 +179,7 @@ import { api } from '@/api'
 import {
   amount,
   maxDecimalDigitsCount,
-  noMoreThanAvailableOnBalance,
+  lessThenMax,
   required,
 } from '@validators'
 import { MathUtil } from '@/js/utils/math.util'
@@ -249,7 +249,7 @@ export default {
         amount: {
           required,
           amount,
-          noMoreThanAvailableOnBalance: noMoreThanAvailableOnBalance(
+          noMoreThanAvailableOnBalance: lessThenMax(
             this.form.asset.balance.value
           ),
           maxDecimalDigitsCount: maxDecimalDigitsCount(